rename Constants.cs

+++ Elegant.DAL/Storages/DbProductsStorage.cs
@@ -28,30 +28,41 @@ public class DbProductsStorage : IProductsStorage
     public void Update(Product product)
     {
         var productInStorage = GetById(product.Id);
-        productInStorage.Name = product.Name;
-        productInStorage.Cost = product.Cost;
-        productInStorage.Description = product.Description;
-        productInStorage.ImagePath = product.ImagePath;
+        if (productInStorage != null)
+        {
+            productInStorage.Name = product.Name;
+            productInStorage.Cost = product.Cost;
+            productInStorage.Description = product.Description;
+            productInStorage.ImagePath = product.ImagePath;
+        }
+
         _dbContext.SaveChanges();
     }
 
     public List<Product> Search(string name)
     {
-        if (string.IsNullOrWhiteSpace(name))
-            return new List<Product>();
-        return _dbContext.Products.Where(product => product.Name.ToLower().StartsWith(name.ToLower())).ToList();
+        return string.IsNullOrWhiteSpace(name)
+            ? new List<Product>()
+            : _dbContext.Products.Where(product => product.Name.ToLower().StartsWith(name.ToLower())).ToList();
     }
 
     public void Remove(Guid productId)
     {
         var product = GetById(productId);
+        if (product == null)
+        {
+            return;
+        }
+
         _dbContext.Products.Remove(product);
         _dbContext.SaveChanges();
     }
 
-    public Product GetById(Guid id)
+    public Product? GetById(Guid id)
     {
-        return _dbContext.Products.Include(product => product.CartItems).Include(product => product.ImageItems)
+        return _dbContext.Products
+            .Include(product => product.CartItems)
+            .Include(product => product.ImageItems)
             .FirstOrDefault(pr => pr.Id == id);
     }
 }
